# docs/swagger.yaml
basePath: /
definitions:
...
  web.IndikatorPokinOpdTacticalCreateRequest:
    properties:
      indikator:
        type: string
    required:
    - indikator
    type: object
  web.IndikatorPokinOpdTacticalResponse:
    properties:
      id:
        type: integer
      indikator:
        type: string
    type: object
  web.IndikatorPokinOpdTacticalUpdateRequest:
    properties:
      id:
        type: integer
      indikator:
        type: string
    required:
    - id
    type: object
...
  web.PohonKinerjaIndikatorResponse:
    properties:
      id:
        type: integer
      indikator:
        type: string
      targets:
        items:
          $ref: '#/definitions/web.PohonKinerjaTargetResponse'
        type: array
    type: object
  web.PohonKinerjaResponse:
    properties:
      childs:
        items:
          $ref: '#/definitions/web.PokinOpdStrategicResponse'
        type: array
      kode_opd:
        type: string
      nama_opd:
        type: string
      tahun:
        type: integer
      tujuan_opd:
        items:
          $ref: '#/definitions/web.PohonKinerjaTujuanResponse'
        type: array
    type: object
  web.PohonKinerjaTargetResponse:
    properties:
      id:
        type: integer
      satuan:
        type: string
      target:
        type: integer
    type: object
  web.PohonKinerjaTujuanResponse:
    properties:
      bidang_urusan:
        type: string
      id:
        type: integer
      indikator:
        items:
          $ref: '#/definitions/web.PohonKinerjaIndikatorResponse'
        type: array
      kode_opd:
        type: string
      tahun_akhir_periode:
        type: integer
      tahun_awal_periode:
        type: integer
      tujuan:
        type: string
    type: object
...
  web.PokinOpdStrategicIndikatorResponse:
    properties:
      id_indikator:
        type: integer
      nama_indikator:
        type: string
      targets:
        items:
          $ref: '#/definitions/web.PokinOpdStrategicTargetResponse'
        type: array
    type: object
  web.PokinOpdStrategicResponse:
    properties:
      id:
        type: integer
      indikator:
        items:
          $ref: '#/definitions/web.PokinOpdStrategicIndikatorResponse'
        type: array
      jenis_pohon:
        type: string
      jumlah_review:
        type: integer
      keterangan:
        type: string
      kode_opd:
        type: string
      level_pohon:
        type: integer
      nama_opd:
        type: string
      nama_pohon:
        type: string
      parent:
        type: integer
      pelaksana:
        type: string
      status:
        type: string
      tahun:
        type: integer
      updated_by:
        type: string
    type: object
  web.PokinOpdStrategicTargetResponse:
    properties:
      id_target:
        type: integer
      indikator_id:
        type: integer
      satuan:
        type: string
      target:
        type: integer
    type: object
...
  web.WebResponse:
    description: Web Response
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
host: ${PROD_HOSTNAME}
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API For Alur Kerja Services
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Alur Kerja Service API
  version: "1.0"
paths:
...
  /indikator-pokin-opd-tacticals:
    get:
      consumes:
      - application/json
      description: Get list of all Indikator Pokin OPD Tactical
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.IndikatorPokinOpdTacticalResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: List All Indikator Pokin OPD Tactical
      tags:
      - Indikator Pokin OPD Tactical
    post:
      consumes:
      - application/json
      description: Create new Indikator Pokin OPD Tactical
      parameters:
      - description: Indikator Pokin OPD Tactical Create Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.IndikatorPokinOpdTacticalCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.IndikatorPokinOpdTacticalResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Create Indikator Pokin OPD Tactical
      tags:
      - Indikator Pokin OPD Tactical
  /indikator-pokin-opd-tacticals/{id}:
    delete:
      consumes:
      - application/json
      description: Delete existing Indikator Pokin OPD Tactical by ID
      parameters:
      - description: Indikator Pokin OPD Tactical ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Delete Indikator Pokin OPD Tactical
      tags:
      - Indikator Pokin OPD Tactical
    get:
      consumes:
      - application/json
      description: Get Indikator Pokin OPD Tactical detail by ID
      parameters:
      - description: Indikator Pokin OPD Tactical ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.IndikatorPokinOpdTacticalResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get Indikator Pokin OPD Tactical by ID
      tags:
      - Indikator Pokin OPD Tactical
    put:
      consumes:
      - application/json
      description: Update existing Indikator Pokin OPD Tactical by ID
      parameters:
      - description: Indikator Pokin OPD Tactical ID
        in: path
        name: id
        required: true
        type: integer
      - description: Indikator Pokin OPD Tactical Update Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.IndikatorPokinOpdTacticalUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.IndikatorPokinOpdTacticalResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Update Indikator Pokin OPD Tactical
      tags:
      - Indikator Pokin OPD Tactical
...
  /pohon-kinerja-opds/{kode_opd}/{tahun}:
    get:
      consumes:
      - application/json
      description: Get pohon kinerja opd by kode opd and tahun
      parameters:
      - description: Kode OPD
        in: path
        name: kode_opd
        required: true
        type: string
      - description: Tahun
        in: path
        name: tahun
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.PohonKinerjaResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get Pohon Kinerja Opd
      tags:
      - Pohon Kinerja Opd
...
schemes:
- http
- https
swagger: "2.0"